🔨refactor: quiz_page.dartに定義しているオブジェクトを別ファイルに切り分け
