+++ lib/constants/box_shadow.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/material.dart';
+
+const List<BoxShadow> boxShadow = [
+  BoxShadow(
+    color: Colors.black26,
+    offset: Offset(0, 2),
+    blurRadius: 4.0,
+  ),
+];
