+++ lib/pages/quiz_page.dart
@@ -1,24 +1,15 @@
-import 'dart:math';
-
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:html_character_entities/html_character_entities.dart';
 
-import '../enums/difficulty.dart';
 import '../models/failure/failure.dart';
 import '../models/question/question.dart';
 import '../providers/quiz/quiz_provider.dart';
 import '../providers/quiz/quiz_state.dart';
-import '../repositories/quiz/quiz_repository.dart';
-
-final quizQuestionsProvider = FutureProvider.autoDispose<List<Question>>(
-  (ref) => ref.watch(quizRepositoryProvider).getQuestions(
-        numQuestions: 5,
-        categoryId: Random().nextInt(24) + 9,
-        difficulty: Difficulty.any,
-      ),
-);
+import '../widgets/custom_button.dart';
+import '../widgets/quiz_error.dart';
+import '../widgets/quiz_questions.dart';
+import '../widgets/quiz_results.dart';
 
 class QuizPage extends HookConsumerWidget {
   const QuizPage({Key? key}) : super(key: key);
@@ -84,7 +75,7 @@ class QuizPage extends HookConsumerWidget {
     if (questions.isEmpty) {
       return QuizError(
         widgetRef: ref,
-        message: 'No questions found.',
+        message: 'クイズをうまく取得できませんでした🥹',
       );
     }
     final quizState = ref.watch(quizStateNotifierProvider);
@@ -100,310 +91,5 @@ class QuizPage extends HookConsumerWidget {
             questions: questions,
             widgetRef: ref,
           );
-  }
-}
-
-class QuizError extends StatelessWidget {
-  const QuizError({
-    Key? key,
-    required this.message,
-    required this.widgetRef,
-  }) : super(key: key);
-
-  final String message;
-  final WidgetRef widgetRef;
-
-  @override
-  Widget build(BuildContext context) {
-    return Center(
-      child: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: [
-          Text(
-            message,
-            style: const TextStyle(
-              color: Colors.white,
-              fontSize: 20.0,
-            ),
-          ),
-          const SizedBox(height: 20.0),
-          CustomButton(
-            title: 'Retry',
-            onTap: () => widgetRef.refresh(quizRepositoryProvider),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-const List<BoxShadow> boxShadow = [
-  BoxShadow(
-    color: Colors.black26,
-    offset: Offset(0, 2),
-    blurRadius: 4.0,
-  ),
-];
-
-class CustomButton extends StatelessWidget {
-  const CustomButton({
-    Key? key,
-    required this.title,
-    required this.onTap,
-  }) : super(key: key);
-
-  final String title;
-  final VoidCallback onTap;
-
-  @override
-  Widget build(BuildContext context) {
-    return GestureDetector(
-      onTap: onTap,
-      child: Container(
-        margin: const EdgeInsets.all(20.0),
-        height: 50.0,
-        width: double.infinity,
-        decoration: BoxDecoration(
-          color: Colors.yellow[700],
-          boxShadow: boxShadow,
-          borderRadius: BorderRadius.circular(25.0),
-        ),
-        alignment: Alignment.center,
-        child: Text(
-          title,
-          style: const TextStyle(
-            fontSize: 18.0,
-            fontWeight: FontWeight.w600,
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class QuizResults extends StatelessWidget {
-  const QuizResults({
-    Key? key,
-    required this.state,
-    required this.questions,
-    required this.widgetRef,
-  }) : super(key: key);
-
-  final QuizState state;
-  final List<Question> questions;
-  final WidgetRef widgetRef;
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      mainAxisAlignment: MainAxisAlignment.center,
-      crossAxisAlignment: CrossAxisAlignment.stretch,
-      children: [
-        Text(
-          '${state.correct.length} / ${questions.length}',
-          style: const TextStyle(
-            color: Colors.white,
-            fontSize: 60.0,
-            fontWeight: FontWeight.w600,
-          ),
-          textAlign: TextAlign.center,
-        ),
-        const Text(
-          'CORRECT',
-          style: TextStyle(
-            color: Colors.white,
-            fontSize: 48.0,
-            fontWeight: FontWeight.bold,
-          ),
-          textAlign: TextAlign.center,
-        ),
-        const SizedBox(height: 40.0),
-        CustomButton(
-          title: 'New Quiz',
-          onTap: () {
-            widgetRef.refresh(quizRepositoryProvider);
-            widgetRef.read(quizStateNotifierProvider.notifier).reset();
-          },
-        ),
-      ],
-    );
-  }
-}
-
-class QuizQuestions extends StatelessWidget {
-  const QuizQuestions({
-    Key? key,
-    required this.state,
-    required this.questions,
-    required this.pageController,
-    required this.widgetRef,
-  }) : super(key: key);
-
-  final PageController pageController;
-  final QuizState state;
-  final List<Question> questions;
-  final WidgetRef widgetRef;
-
-  @override
-  Widget build(BuildContext context) {
-    return PageView.builder(
-      controller: pageController,
-      physics: const NeverScrollableScrollPhysics(),
-      itemCount: questions.length,
-      itemBuilder: (BuildContext context, int index) {
-        final question = questions[index];
-        return Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            Text(
-              'Question ${index + 1} of ${questions.length}',
-              style: const TextStyle(
-                color: Colors.white,
-                fontSize: 24.0,
-                fontWeight: FontWeight.bold,
-              ),
-            ),
-            Padding(
-              padding: const EdgeInsets.fromLTRB(20.0, 16.0, 20.0, 12.0),
-              child: Text(
-                HtmlCharacterEntities.decode(question.question),
-                style: const TextStyle(
-                  color: Colors.white,
-                  fontSize: 28.0,
-                  fontWeight: FontWeight.w500,
-                ),
-              ),
-            ),
-            Divider(
-              color: Colors.grey[200],
-              height: 32.0,
-              thickness: 2.0,
-              indent: 20.0,
-              endIndent: 20.0,
-            ),
-            Column(
-              children: question.answers
-                  .map(
-                    (e) => AnswerCard(
-                      answer: e,
-                      isSelected: e == state.selectedAnswer,
-                      isCorrect: e == question.correctAnswer,
-                      isDisplayingAnswer: state.answered,
-                      onTap: () => widgetRef
-                          .read(quizStateNotifierProvider.notifier)
-                          .submitAnswer(question, e),
-                    ),
-                  )
-                  .toList(),
-            ),
-          ],
-        );
-      },
-    );
-  }
-}
-
-class AnswerCard extends StatelessWidget {
-  const AnswerCard({
-    Key? key,
-    required this.answer,
-    required this.isSelected,
-    required this.isCorrect,
-    required this.isDisplayingAnswer,
-    required this.onTap,
-  }) : super(key: key);
-
-  final String answer;
-  final bool isSelected;
-  final bool isCorrect;
-  final bool isDisplayingAnswer;
-  final VoidCallback onTap;
-
-  @override
-  Widget build(BuildContext context) {
-    return GestureDetector(
-      onTap: onTap,
-      child: Container(
-        margin: const EdgeInsets.symmetric(
-          vertical: 12.0,
-          horizontal: 20.0,
-        ),
-        padding: const EdgeInsets.symmetric(
-          vertical: 12.0,
-          horizontal: 20.0,
-        ),
-        width: double.infinity,
-        decoration: BoxDecoration(
-          color: Colors.white,
-          boxShadow: boxShadow,
-          border: Border.all(
-            color: isDisplayingAnswer
-                ? isCorrect
-                    ? Colors.green
-                    : isSelected
-                        ? Colors.red
-                        : Colors.white
-                : Colors.white,
-            width: 4.0,
-          ),
-          borderRadius: BorderRadius.circular(100.0),
-        ),
-        child: Row(
-          mainAxisAlignment: MainAxisAlignment.spaceBetween,
-          children: [
-            Flexible(
-              child: Text(
-                HtmlCharacterEntities.decode(answer),
-                style: TextStyle(
-                  color: Colors.black,
-                  fontSize: 16.0,
-                  fontWeight: isDisplayingAnswer && isCorrect
-                      ? FontWeight.bold
-                      : FontWeight.w400,
-                ),
-              ),
-            ),
-            if (isDisplayingAnswer)
-              isCorrect
-                  ? const CircularIcon(icon: Icons.check, color: Colors.green)
-                  : isSelected
-                      ? const CircularIcon(
-                          icon: Icons.close,
-                          color: Colors.red,
-                        )
-                      : const SizedBox.shrink()
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-class CircularIcon extends StatelessWidget {
-  final IconData icon;
-  final Color color;
-
-  const CircularIcon({
-    Key? key,
-    required this.icon,
-    required this.color,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      height: 24.0,
-      width: 24.0,
-      decoration: BoxDecoration(
-        color: color,
-        shape: BoxShape.circle,
-        boxShadow: boxShadow,
-      ),
-      child: Icon(
-        icon,
-        color: Colors.white,
-        size: 16.0,
-      ),
-    );
   }
 }
