+++ lib/widgets/circular_icon.dart
@@ -1,0 +1,32 @@
+import 'package:flutter/material.dart';
+
+import '../constants/box_shadow.dart';
+
+class CircularIcon extends StatelessWidget {
+  final IconData icon;
+  final Color color;
+
+  const CircularIcon({
+    Key? key,
+    required this.icon,
+    required this.color,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      height: 24.0,
+      width: 24.0,
+      decoration: BoxDecoration(
+        color: color,
+        shape: BoxShape.circle,
+        boxShadow: boxShadow,
+      ),
+      child: Icon(
+        icon,
+        color: Colors.white,
+        size: 16.0,
+      ),
+    );
+  }
+}
