+++ lib/widgets/custom_button.dart
@@ -1,0 +1,39 @@
+import 'package:flutter/material.dart';
+
+import '../constants/box_shadow.dart';
+
+class CustomButton extends StatelessWidget {
+  const CustomButton({
+    Key? key,
+    required this.title,
+    required this.onTap,
+  }) : super(key: key);
+
+  final String title;
+  final VoidCallback onTap;
+
+  @override
+  Widget build(BuildContext context) {
+    return GestureDetector(
+      onTap: onTap,
+      child: Container(
+        margin: const EdgeInsets.all(20.0),
+        height: 50.0,
+        width: double.infinity,
+        decoration: BoxDecoration(
+          color: Colors.yellow[700],
+          boxShadow: boxShadow,
+          borderRadius: BorderRadius.circular(25.0),
+        ),
+        alignment: Alignment.center,
+        child: Text(
+          title,
+          style: const TextStyle(
+            fontSize: 18.0,
+            fontWeight: FontWeight.w600,
+          ),
+        ),
+      ),
+    );
+  }
+}
