+++ lib/widgets/quiz_questions.dart
@@ -1,0 +1,81 @@
+import 'package:flutter/material.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'package:html_character_entities/html_character_entities.dart';
+
+import '../models/question/question.dart';
+import '../providers/quiz/quiz_provider.dart';
+import '../providers/quiz/quiz_state.dart';
+import 'answer_card.dart';
+
+class QuizQuestions extends StatelessWidget {
+  const QuizQuestions({
+    Key? key,
+    required this.state,
+    required this.questions,
+    required this.pageController,
+    required this.widgetRef,
+  }) : super(key: key);
+
+  final PageController pageController;
+  final QuizState state;
+  final List<Question> questions;
+  final WidgetRef widgetRef;
+
+  @override
+  Widget build(BuildContext context) {
+    return PageView.builder(
+      controller: pageController,
+      physics: const NeverScrollableScrollPhysics(),
+      itemCount: questions.length,
+      itemBuilder: (BuildContext context, int index) {
+        final question = questions[index];
+        return Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            Text(
+              'Question ${index + 1} of ${questions.length} 🥳',
+              style: const TextStyle(
+                color: Colors.white,
+                fontSize: 24.0,
+                fontWeight: FontWeight.bold,
+              ),
+            ),
+            Padding(
+              padding: const EdgeInsets.fromLTRB(20.0, 16.0, 20.0, 12.0),
+              child: Text(
+                HtmlCharacterEntities.decode(question.question),
+                style: const TextStyle(
+                  color: Colors.white,
+                  fontSize: 28.0,
+                  fontWeight: FontWeight.w500,
+                ),
+              ),
+            ),
+            Divider(
+              color: Colors.grey[200],
+              height: 32.0,
+              thickness: 2.0,
+              indent: 20.0,
+              endIndent: 20.0,
+            ),
+            Column(
+              children: question.answers
+                  .map(
+                    (e) => AnswerCard(
+                      answer: e,
+                      isSelected: e == state.selectedAnswer,
+                      isCorrect: e == question.correctAnswer,
+                      isDisplayingAnswer: state.answered,
+                      onTap: () => widgetRef
+                          .read(quizStateNotifierProvider.notifier)
+                          .submitAnswer(question, e),
+                    ),
+                  )
+                  .toList(),
+            ),
+          ],
+        );
+      },
+    );
+  }
+}
